benchmarks/MOABB/extra-requirements.txt
mne
moabb
orion[profet, bohb]
scikit-learn
torch_geometric
torchinfo